engine/src/programs/python/shader-embedder/pyproject.toml
[project]
name = "shader-embedder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pybars3>=0.9.7",
    "pydantic>=2.12.5",
    "tap>=0.2",
]
